display: Use DO_NOT_DIRECTLY_USE_ME_IMPLEMENT_META_INTERFACE
Fixes error of this type with the the files changed

    error: static_assert failed due to requirement 'int
ernal::allowedManualInterface("android.display.IQHDMIClient")'
"b/64223827: Manually written binder interfaces are considered e
rror prone and frequently have bugs. The preferred way to add
interfaces is to define an .aidl file to auto-generate the interf
ace. If an interface must be manually written, add its name to the
whitelist."